requirements.txt
streamlit
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pydrive2
Pillow
requests
